refactor: added userTypeToTitleCase instead of hardcoded values for user types
